test: merge ReturnsEmptyValuesFutureTest into ReturnsEmptyValuesTest
